Updated app launch script to make single instance detection work with High Sierra (closes #813).
Something has changed with detecting multiple instances of applications
that use a shell script as its executable between Sierra and High Sierra
versions of MacOS, as the previously used mechanism allowed a second instance
to be launched in a "broken" state (with a bouncing Dock icon), which
stayed completely unresponsive. Adding `exec`, which replaces the shell
instance with the started executable, appears to fix the issue.