Updated to call `onEditorClose` before document removal.
This is a partial reversal of the changes in 0affb36f and 429adca3
that changes the semantics of onEditorClose to be called after document removal,
which affected plugins that depend on that behavior.

Since it's difficult to document and a bit inconsistent with other
editor methods, this behavior has been reversed.