Updated to allow `StartStyling` to work with deprecated parameters (closes #994).
wxwidgets 3.1.2+ includes a version of Scintilla (wxSTC) that deprecated
the second parameter of StartStyling and throws an assert when it's set
to a non-zero value. Unfortunately, this conflicts with the usage for
earlier versions and because Scintilla version number is not available
in wxlua (this is fixed in pkulchenko/wxlua#33), there is no direct
check to apply when the change was made.

As a workaround, the check is made for `wxSTC_INDIC0_MASK`, which was
removed in the same commit as `StartStyling` was updated.