Don't background IDE job when launched from shell wrapper
This is a shell anti-pattern and hinders integration into normal Linux
desktop usage. Only daemon launchers should background a job, and even
then only on request.

Instead we actually want to `exec` it, which has the effect of cleaning
up the wrapper script processes, making sure all the signals are
connected to the _actual_ foreground process, and passing exit codes and
such through when all is said and done.