Allow environment variable to set the Lua executable
For some distros (e.g. Arch Linux) packaging binaries compiled elsewhere
is a big no-no, and the more so when they are duplicates of something
the system has anyway. The current Arch Linux packaging for example has
to remove the generated binaries from the installation, then patch this
wrapper script to use the system Lua executable.

This change will make that process easier. First it will be possible to
run with a different Lua executable by setting an environment variable:

    LUA_EXECUTABLE=/usr/bin/lua zbstudio

Second, the coding style in the script will make it much easier to patch
to make the that choice stick.