Fixed processing of recovery record that may not have modification timestamp (fixes #859).
This happens under rare circumstances when an opened file is removed on
disk, which is then detected by the IDE (which removes the timestamp
record). If the editor content is then modified and the file is restored
on disk and the IDE is restarted (saving the modified content in its
recovery record), this may trigger the error that will prevent the IDE
from properly launching.

May also fix #1085.